gcc.mk: Simplify version-choice comments
Our documented plan is to choose a version that supports the standard,
and to choose limited versions as found in NetBSD.  When the choice
straightforwardly follows upstream's documented requirements, and when
it doesn't meet upstream's documented minimum, the "limited versions"
guideline does not come in to play, so don't mention it.  (The limited
versions guideline would make sense to mention when we choose a higher
version than is required to avoid proliferation, but this does not
seem to be happening at the moment.)

Move comment about 4.8 in CentOS 7 to 'we tend' paragraph.

(This is a comment-only change.)