textproc/enchant2: Rewrite DESCR from upstream and fold in MESSAGE
Upstream's text is less sales-y and easier to follow now.  Adopt it,
and prune things not supported by pkgsrc (various spell checkers) and
information not belonging in DESCR (languages used).  Note in DESCR
that this package does not install language-specific dictionaries in
dependent spell checkers, to avoid the description of this package
leaving the wrong impression for those who assume that would make
sense.

Drop MESSAGE, having folded into DESCR a caution that an API library
that depends on a spell checker does not install the language-specific
dictionary that that user might have in mind.