opa: Update to 0.66.0
Changes:
v0.66.0
This release contains a mix of features, performance improvements, and bugfixes.

### Improved Test Reports

The `opa test` command now includes a new `--var-values` flag that
enriches reporting of failed tests with the values and locations for
variables in the failing expression.

### Reading stdin in `opa exec`

The `opa exec` command now supports reading `input` documents from
stdin with the `--stdin-input` (`-I`) flag.


v0.65.0
This release contains a mix of features and bugfixes.