devel/py-test-textual-snapshot: import py-test-textual-snapshot-0.4.0
Snapshot testing for Textual apps.

A pytest-textual-snapshot test saves an SVG screenshot of a running
Textual app to disk.  The next time the test runs, it takes another
screenshot and compares it to the saved one.  If the new screenshot
differs from the old one, the test fails.  This is a convenient
way to quickly and automatically detect visual regressions in your
applications.