geography/mapserver: Update to 8.2.1
This package was last updated in 2012, and has been failing to build
because it uses a withdrawn proj API.  This update is rougher than our
normal standards, but brings the package to the current release and
successfully builds for me.

Packaging changes:

  - convert to cmake

  - drop all patches because they are to files which no longer exist and
    are either
      + clearly not applicable (to configure), or
      + lack patch comments and upstream bug URLs

  - depend on protobuf-c because the build complained

It remains to

  - align the list of bl3 files with the upstream build docs

  - set cmake flags harder, to remediate cmake's bug of finding things
    that aren't bl3d

  - actually test

Upstream publishes changelogs which are vast and therefore not really
helpful, and does not publish NEWS.  Users are cautioned to treat this
like a new package.  Given the lack of complaints, there are likely
few.