py-websockets: update to 13.0.
13.0

August 20, 2024

Backwards-incompatible changes

websockets 13.0 requires Python ≥ 3.8.

websockets 12.0 is the last version supporting Python 3.7.

Receiving the request path in the second parameter of connection handlers is deprecated.

If you implemented the connection handler of a server as:

async def handler(request, path):
  ...

You should switch to the pattern recommended since version 10.1:

async def handler(request):
    path = request.path  # only if handler() uses the path argument
    ...

The ssl_context argument of connect() and serve() in the threading implementation is renamed to ssl.

This aligns the API of the threading implementation with the asyncio implementation.

For backwards compatibility, ssl_context is still supported.

The WebSocketServer class in the threading implementation is renamed to Server.

This change should be transparent because this class shouldn’t be instantiated directly; serve() returns an instance.

Regardless, an alias provides backwards compatibility.

New features

websockets 11.0 introduces a new asyncio implementation.

This new implementation is intended to be a drop-in replacement for the current implementation. It will become the default in a future release.

Please try it and report any issue that you encounter! The upgrade guide explains everything you need to know about the upgrade process.

    Validated compatibility with Python 3.12 and 3.13.

    Added environment variables to configure debug logs, the Server and User-Agent headers, as well as security limits.

    If you were monkey-patching constants, be aware that they were renamed, which will break your configuration. You must switch to the environment variables.

Improvements

    The error message in server logs when a header is too long is more explicit.

Bug fixes

    Fixed a bug in the threading implementation that could prevent the program from exiting when a connection wasn’t closed properly.

    Redirecting from a ws:// URI to a wss:// URI now works.

    broadcast(raise_exceptions=True) no longer crashes when there isn’t any exception.