py-posting: update to 1.13.0.
1.13.0

What's Changed

With this release, Posting will watch loaded environment files and
refresh the UI to reflect changes in them. There's some new config
for hiding the collection browser on startup too, and a bunch of
smaller changes. See the log below for more info.

Added

    New collection_browser.show_on_startup config to control whether the collection browser is shown on startup.
    Watch for changes to loaded dotenv files and reload UI elements that depend on them when they change.

Changed

    Upgraded all dependencies.
    Remove pydantic-settings crash workaround on empty config files.
    Renaming App.maximized as it now clashes with a Textual concept.
    Removed "using default collection" message from startup.

Fixed

    Fixed crash while rendering error message on timeout.

1.12.3

What's Changed

    Upgrade textual by @darrenburns in #94
    Catch possible pyperclip exception if no clipboard mechanism installed by @seapagan in #98

1.12.2

Note: 1.12.2 was yanked. It's equivalent to 1.12.1.