add security/liboqs version 0.10.1
liboqs is an open source C library for quantum-safe cryptographic
algorithms.  It provides a collection of open-source implementations of
quantum-safe key encapsulation mechanism (KEM) and digital signature
algorithms, a common API for these algorithms, and a test harness and
benchmarking routines.

`liboqs` is part of the Open Quantum Safe (OQS) project.

This package is largely based on work done by logan@cyberstorm.mu in
pkgsrc-wip with minor modifications by myself.