mk: Add mktool support to digest.
mktool is an optional external tool, available from pkgtools/mktool or
"cargo install mktool", that provides a "mktool digest" command that is
compatible with pkgtools/digest.

After installing, set TOOLS_PLATFORM.mktool=/path/to/mktool to enable it.

The benefit of using mktool for the DIGEST tool is that it eliminates the
dependency on pkgtools/digest for almost every package.  This has a
particular impact in bulk builds, where dependencies are installed fresh for
every single package build.

Currently "mktool digest" only implements the algorithms that are actually
used (or have been used in the past) by pkgsrc, though adding support for
e.g. tiger and whirlpool would be trivial.