xcmsdb: update to 1.0.7.
This release fixes some compiler warnings and includes other minor code
cleanups.

For those building for 32-bit platforms, it also enables use of the
"large file" APIs - though no file processed by this program should ever
be anywhere near 2GB in size, it is a prerequisite for post-y2038 support.
Since this release was generated using the new GNU autoconf 2.72, this
also adds a --enable-year2038 configure flag which may allow it to
work with files whose timestamps are later than January 19, 2038, but
this has not been tested.

Alan Coopersmith (10):
      gitlab CI: stop requiring Signed-off-by in commits
      ParseVisualOptions: eliminate unneccessary strcpy & strcat
      Variable scope reduction as suggested by cppcheck
      Declare DefineEntry tables of string to id mappings as const
      Correct comment describing DefineToStr
      SCScrnClassStringOf: always return a non-NULL string
      ProcessIProfile: Avoid NULL dereference if strtok doesn't find a match
      ProcessColorimetric: limit string size written by scanf()
      configure: Use AC_SYS_LARGEFILE to enable large file support
      xcmsdb 1.0.7