xkbprint: update to 1.0.7.
This release provides some code cleanups and compiler warning fixes.

For those building for 32-bit platforms, it also enables use of the
"large file" APIs - though no file processed by this program should ever
be anywhere near 2GB in size, it is a prerequisite for post-y2038 support.
Since this release was generated using the new GNU autoconf 2.72, this
also adds a --enable-year2038 configure flag which may allow it to
work with files whose timestamps are later than January 19, 2038, but
this has not been tested.

Alan Coopersmith (7):
      unifdef -Usgi
      configure: Use AC_SYS_LARGEFILE to enable large file support
      ClearFontStuff: avoid NULL pointer dereference
      FindKeysymsByName: avoid NULL pointer dereference
      Variable scope reductions
      Use C99 struct initializers
      xkbprint 1.0.7