mk: Add ARGMAX_REQ support for ARG_MAX limits.
The wip/grafana package now contains so many distfiles that any command
that operates on them now overflows ARG_MAX on NetBSD (262144).  This
causes problems both in how bmake handles executing shell commands, as
well as the checksum.awk script that executes digest on them.

Modify checksum.awk to support reading distfiles from stdin, and process
files in batches if a digest command is over 64K in length.  This value
strikes a good balance between performance and ensuring even the most
stringent modern ARG_MAX limit is avoided.

Introduce a new package variable ARGMAX_REQ that if set to a value higher
than ARG_MAX, and if we are unable to detect a bmake that has its own fix
for this problem, will use alternate targets that use temporary files for
input.

With this change, and ARGMAX_REQ=524288 set in wip/grafana, NetBSD's
native make is now able to both create and check its distinfo.

As an added benefit, on systems that do not breach the limit and are
running a newer bmake, the distinfo and makesum targets no longer use a
temporary input file.