tex-newcomputermodern{,-doc}: update to 6.0.0
This is a major release of NewComputerModern.

The most notable change is the introduction of a Bold Math font
which completely covers the Unicode Math blocks and even beyond
that (with the same coverage as Regular and Book).

Other improvements:

NewCM mathbb symbols (in ss03) have been reworked and they now
match the design of ComputerModern fonts (the AMSbb remain the
default)

AMSbb and NewCMbb are now provided in SmallCaps for special
applications (see the documentation) in ss05

An additional operator similar to \sum and \int has been added
to the Math fonts. Namely now it is possible to convolute many
function in the usual shorthand notation mathematicians use.
The new operator is called: \convolution

Added one more radical for smother change from one to another

Added Currency Symbols font in all text fonts (except Uncials)

Fixed several typos in Lookup tables thanks to Eric Biedert

Fixed CapHeight parameter in OS2 tables for some of the fonts
thanks to Max Kortge