lang/python: Update comment about needed build systems
Setuptools now can create wheels (per py-wheel upstream README, and
actually trying it).

Clarify discussion of PEP-518: While the default is setuptools and
wheel if pyproject.toml doesn't have a build section, the actual
default expectation is now only (non-ancient) setuptools.

Give an example of the needed TOOL_DEPENDS; the point of
lang/python/foo.mk is to be easy to use and to abstract the repetitive
details.